# No need to set OMP num_threads