@martin-frbg 2022-02-25
Reason given for this change in #2775 (@Guobing-Chen) was
"In current code, no matter what number of threads specified, all
available CPU count is used when invoking OMP, which leads to very bad
performance if the workload is small while all available CPUs are big.
Lots of time are wasted on inter-thread sync. Fix this issue by really
using the number specified by the variable 'num' from calling API."
so I am a bit sceptical you may just be comparing different situations/workloads