@brada4 2022-03-03
That is rather ancient benchmark script, please re-check with benchmarks/scripts/R/deig.R
Namely it permits gc() in metered section.
How does pthread version perform?