@kangshan1157 2022-03-03
> That is rather ancient benchmark script, please re-check with benchmarks/scripts/R/deig.R Namely it permits gc() in metered section. How does pthread version perform?

I tried "benchmarks/scripts/R/deig.R" and here are the openMP version (USE_THREAD=1 USE_OPENMP=1) results.
Without the patch
      SIZE             Flops                   Time
           128x128 :    1927.93 MFlops   0.029000 sec
           256x256 :     570.51 MFlops   0.784000 sec
           384x384 :     721.25 MFlops   2.093000 sec
           512x512 :     378.53 MFlops   9.453000 sec
           640x640 :     409.15 MFlops  17.081000 sec
           768x768 :     475.10 MFlops  25.419000 sec
           896x896 :     582.22 MFlops  32.938000 sec
         1024x1024 :     679.14 MFlops  42.150000 sec
         1152x1152 :     784.43 MFlops  51.959000 sec
         1280x1280 :     902.08 MFlops  61.979000 sec
         1408x1408 :    1038.00 MFlops  71.692000 sec
         1536x1536 :    1170.07 MFlops  82.570000 sec
         1664x1664 :    1223.62 MFlops 100.386000 sec
         1792x1792 :    1371.79 MFlops 111.837000 sec
         1920x1920 :    1530.71 MFlops 123.274000 sec
         2048x2048 :    1683.94 MFlops 135.995000 sec

With the patch:
       SIZE             Flops                   Time
           128x128 :    1189.58 MFlops   0.047000 sec
           256x256 :    4758.30 MFlops   0.094000 sec
           384x384 :    7987.15 MFlops   0.189000 sec
           512x512 :    8321.50 MFlops   0.430000 sec
           640x640 :   11805.34 MFlops   0.592000 sec
           768x768 :   14946.26 MFlops   0.808000 sec
           896x896 :   18801.13 MFlops   1.020000 sec
         1024x1024 :   15634.06 MFlops   1.831000 sec
         1152x1152 :   17147.01 MFlops   2.377000 sec
         1280x1280 :   15217.77 MFlops   3.674000 sec
         1408x1408 :   24641.16 MFlops   3.020000 sec
         1536x1536 :   32376.88 MFlops   2.984000 sec
         1664x1664 :   22427.32 MFlops   5.477000 sec
         1792x1792 :   29691.74 MFlops   5.167000 sec
         1920x1920 :   27672.17 MFlops   6.819000 sec
         2048x2048 :   37297.66 MFlops   6.140000 sec