@brada4 2022-03-03
>  640x640 : 409.15 MFlops 17.081000 sec
>  640x640 : 11805.34 MFlops 0.592000 sec

Impressive indeed.

How it compares to pthread version, i.e building without USE_OPENMP parameter? Not picking on you, just curious, over course of day I will measure myself.