@kangshan1157 2022-03-19
@martin-frbg Except the benchmarks/scripts/R/deig.R, what other benchmark do I need to verify. For deig.R this benchmark, it seems it is not stable enough, so I add the loop count to 20 and collect the following data on Ice Lake.
From 128 To 384 Step=128 Loops=20
without patch
  | round 0 | round 0 | round 1 | round 1 | round 2 | round 2 | round 3 | round 3 | round 4 | round 4 | mean | mean | rsd | rsd
-- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | --
SIZE | Flops | Time | Flops | Time | Flops | Time | Flops | Time | Flops | Time | Flops | Time | Flops | Time
128x128 | 2167.06 | 0.516 | 2188.26 | 0.511 | 2129.91 | 0.525 | 2101.88 | 0.532 | 2146.26 | 0.521 | 2146.674 | 0.521 | 1.55% | 1.55%
256x256 | 457.25 | 19.564 | 432.09 | 20.703 | 458.37 | 19.516 | 438.19 | 20.415 | 456.18 | 19.61 | 448.416 | 19.9616 | 2.75% | 2.78%
384x384 | 422.31 | 71.492 | 403.81 | 74.766 | 426.01 | 70.871 | 406.3 | 74.308 | 423.49 | 71.292 | 416.384 | 72.5458 | 2.51% | 2.53%

with patch
  | round 0 | round 0 | round 1 | round 1 | round 2 | round 2 | round 3 | round 3 | round 4 | round 4 | mean | mean | rsd | rsd
-- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | --
SIZE | Flops | Time | Flops | Time | Flops | Time | Flops | Time | Flops | Time | Flops | Time | Flops | Time
128x128 | 2733.99 | 0.409 | 2675.12 | 0.418 | 2463 | 0.454 | 2564.68 | 0.436 | 2733.99 | 0.409 | 2774.69 | 0.403 | 4.25% | 4.84%
256x256 | 5850.63 | 1.529 | 5591.01 | 1.6 | 5305.82 | 1.686 | 5580.54 | 1.603 | 5756.51 | 1.554 | 5893.02 | 1.518 | 3.53% | 3.95%
384x384 | 8569.81 | 3.523 | 8890.29 | 3.396 | 7995.61 | 3.776 | 8521.43 | 3.543 | 8562.52 | 3.526 | 8398.17 | 3.595 | 3.84% | 3.84%

with patch VS without patch
SIZE | Flops
-- | --
128x128 | 1.29x
256x256 | 13.14x
384x384 | 20.17x

Without this path, the larger size, the more time needs to be completed, so I only pickup 3 sizes. The patch can improve this benchmark. I have no idea what else benchmarks I have to verify with the patch.